Big code cleanup
I used IntelliJ IDEA's code analysis feature to help find warnings and bugs, and manually went through fixing them. The only warnings I let IDEA fix automatically were unused imports and redundant casts.
ItemStacks have been converted to use ItemStack.EMPTY instead of null, to conform to new standards.
Annotations have been added to some of the code to help devs determine when null can and can't be used, however they're nowhere near all-encompassing.
Vague fields and variables have been renamed where I saw them, such as i -> itemStack. This excludes math variables that seemed intentionally named, and standardized names like loop control variables.
I fixed a couple of issues in the French lang file, but the non-en_US languages still have some missing translation keys. Probably best for someone who speaks those languages to add the translations, rather than me with Google Translate.
There were some recipes that had invalid json, so I fixed them. This resulted in a few of the recipes changing slightly.
This commit fixes some potential NPEs, though I didn't get them all.

I highly recommend one of the original devs merge this and also run the source through their own IDE's code analysis, since there are still a lot of warnings I couldn't or didn't fix, and a large amount of them could be or turn into bugs.
The ItemStack changes affected some API methods. Mods developed against libVulpes should use builds containing this commit in the future, to prevent them from sending null ItemStacks.
This commit should be used with a sister-commit for Advanced Rocketry, since the ItemStack refactoring started there.

+++ src/main/java/zmaster587/libVulpes/inventory/modules/ModuleScaledImage.java
@@ -74,10 +74,10 @@ public class ModuleScaledImage extends ModuleBase {
 		GlStateManager.color(alpha, alpha, alpha, alpha);
         BufferBuilder buff = Tessellator.getInstance().getBuffer();
         buff.begin(GL11.GL_QUADS, DefaultVertexFormats.POSITION_TEX);
-        buff.pos((double)(x + this.offsetX), (double)(y + this.offsetY + sizeY), (double)0).tex(minX, maxY).endVertex();
-        buff.pos((double)(x + this.offsetX + sizeX), (double)(y + this.offsetY + sizeY), (double)0).tex(maxX, maxY).endVertex();
-        buff.pos((double)(x + this.offsetX + sizeX), (double)(y + this.offsetY), (double)0).tex(maxX, minY).endVertex();
-        buff.pos((double)(x + this.offsetX), (double)(y + this.offsetY), (double)0).tex(minX, minY).endVertex();
+        buff.pos(x + this.offsetX, y + this.offsetY + sizeY, 0).tex(minX, maxY).endVertex();
+        buff.pos(x + this.offsetX + sizeX, y + this.offsetY + sizeY, 0).tex(maxX, maxY).endVertex();
+        buff.pos(x + this.offsetX + sizeX, y + this.offsetY, 0).tex(maxX, minY).endVertex();
+        buff.pos(x + this.offsetX, y + this.offsetY, 0).tex(minX, minY).endVertex();
         Tessellator.getInstance().draw();
         if(alpha < 1f) {
 			GL11.glColor4d(1f, 1f, 1f, 1f);
